// shared/types/FooterType.ts
// types/FooterType.ts
export interface FooterType {
    empowering: string;
    quickLinks: {
      title: string;
      about: string;
      courses: string;
      tutors: string;
      blog: string;
    };
    support: {
      title: string;
      faq: string;
      contact: string;
    };
    contact: {
      title: string;
      address: string;
    };
    rights: string;
  }